+++ pages/nft-details.tsx
@@ -2,7 +2,6 @@ import Image from 'next/image';
 import Link from 'next/link';
 import withTransition from '../components/withTransition';
 import { Button, Loader, Modal } from '../components';
-import { NFTCollection, fetchNFTCollectionMetadata } from '../components/NFTCollection';
 import { NFTMarketplace } from '../utils/nft-marketplace';
 import { ONE_ALPH, prettifyAttoAlphAmount, binToHex, contractIdFromAddress, web3, NodeProvider } from '@alephium/web3'
 import { defaultNodeUrl, marketplaceContractId } from '../configs/nft';
@@ -14,7 +13,7 @@ import { useAlephiumConnectContext } from '@alephium/web3-react';
 import { useRouter } from 'next/router';
 import { useState, useEffect } from 'react';
 import { waitTxConfirmed, shortenName } from '../utils';
-import { NFTCollection as NFTCollectionHelper } from '../utils/nft-collection';
+import { NFTCollectionHelper, NFTCollection, fetchNFTCollectionMetadata } from '../utils/nft-collection';
 import { fetchMintedNFT, NFT } from '../utils/nft';
 
 interface PaymentBodyCmpProps {
