+++ utils/nft-collection.ts
@@ -1,12 +1,44 @@
-import * as web3 from '@alephium/web3'
 import { DeployHelpers } from './deploy-helpers'
-import { NFTOpenCollection, NFTOpenCollectionInstance, NFTPublicSaleCollectionRandom, NFTPublicSaleCollectionRandomInstance, NFTPublicSaleCollectionSequential, NFTPublicSaleCollectionSequentialInstance } from '../artifacts/ts'
-import { CreateOpenCollection, CreatePublicSaleCollectionSequential, MintBatchSequential, MintNextSequential, MintOpenNFT, MintSpecificPublicSaleNFT, WithdrawFromPublicSaleCollection } from '../artifacts/ts/scripts'
-import { DeployContractResult, DUST_AMOUNT, ONE_ALPH } from '@alephium/web3'
+import {
+  NFTOpenCollection,
+  NFTOpenCollectionInstance,
+  NFTOpenCollectionTypes,
+  NFTPublicSaleCollectionRandom,
+  NFTPublicSaleCollectionRandomInstance,
+  NFTPublicSaleCollectionSequential,
+  NFTPublicSaleCollectionSequentialInstance,
+  NFTPublicSaleCollectionSequentialTypes
+} from '../artifacts/ts'
+import {
+  CreateOpenCollection,
+  CreatePublicSaleCollectionSequential,
+  MintBatchSequential,
+  MintNextSequential,
+  MintOpenNFT,
+  MintSpecificPublicSaleNFT,
+  WithdrawFromPublicSaleCollection
+} from '../artifacts/ts/scripts'
+import {
+  web3,
+  DeployContractResult,
+  DUST_AMOUNT,
+  ONE_ALPH,
+  stringToHex,
+  groupOfAddress,
+  addressFromContractId,
+  binToHex,
+  hexToBinUnsafe,
+  hexToString,
+  contractIdFromAddress,
+  subContractId,
+  encodeU256
+} from '@alephium/web3'
 import { nftTemplateId, openCollectionTemplateId, publicSaleCollectionTemplateId } from '../configs/nft'
 import * as blake from 'blakejs'
+import axios from 'axios'
+import { NFT } from './nft'
 
-export class NFTCollection extends DeployHelpers {
+export class NFTCollectionHelper extends DeployHelpers {
   async createOpenCollection(
     collectionUri: string
   ): Promise<DeployContractResult<NFTOpenCollectionInstance>> {
@@ -16,17 +48,17 @@ export class NFTCollection extends DeployHelpers {
       initialFields: {
         openCollectionTemplateId: openCollectionTemplateId,
         nftTemplateId: nftTemplateId,
-        collectionUri: web3.stringToHex(collectionUri),
+        collectionUri: stringToHex(collectionUri),
         collectionOwner: ownerAddress,
         totalSupply: 0n
       },
       attoAlphAmount: ONE_ALPH
     })
-    const groupIndex = web3.groupOfAddress(ownerAddress)
+    const groupIndex = groupOfAddress(ownerAddress)
     const contractId = await calcContractId(result.txId, groupIndex)
     return {
       ...result,
-      contractInstance: NFTOpenCollection.at(web3.addressFromContractId(contractId))
+      contractInstance: NFTOpenCollection.at(addressFromContractId(contractId))
     }
   }
 
@@ -42,8 +74,8 @@ export class NFTCollection extends DeployHelpers {
       {
         initialFields: {
           nftTemplateId: nftTemplateId,
-          collectionUri: web3.stringToHex(collectionUri),
-          nftBaseUri: web3.stringToHex(baseUri),
+          collectionUri: stringToHex(collectionUri),
+          nftBaseUri: stringToHex(baseUri),
           collectionOwner: ownerAddress,
           maxSupply: maxSupply,
           mintPrice: mintPrice,
@@ -67,8 +99,8 @@ export class NFTCollection extends DeployHelpers {
       initialFields: {
         publicSaleCollectionTemplateId: publicSaleCollectionTemplateId,
         nftTemplateId: nftTemplateId,
-        collectionUri: web3.stringToHex(collectionUri),
-        nftBaseUri: web3.stringToHex(baseUri),
+        collectionUri: stringToHex(collectionUri),
+        nftBaseUri: stringToHex(baseUri),
         collectionOwner: ownerAddress,
         maxSupply: maxSupply,
         mintPrice: mintPrice,
@@ -77,11 +109,11 @@ export class NFTCollection extends DeployHelpers {
       },
       attoAlphAmount: ONE_ALPH
     })
-    const groupIndex = web3.groupOfAddress(ownerAddress)
+    const groupIndex = groupOfAddress(ownerAddress)
     const contractId = await calcContractId(result.txId, groupIndex)
     return {
       ...result,
-      contractInstance: NFTPublicSaleCollectionSequential.at(web3.addressFromContractId(contractId))
+      contractInstance: NFTPublicSaleCollectionSequential.at(addressFromContractId(contractId))
     }
   }
 
@@ -125,7 +157,7 @@ export class NFTCollection extends DeployHelpers {
       {
         initialFields: {
           nftCollection: nftCollectionContractId,
-          uri: web3.stringToHex(nftUri)
+          uri: stringToHex(nftUri)
         },
         attoAlphAmount: BigInt(1.1e18)
       }
@@ -169,10 +201,171 @@ export class NFTCollection extends DeployHelpers {
 }
 
 async function calcContractId(txId: string, groupIndex: number) {
-  const nodeProvider = web3.web3.getCurrentNodeProvider()
+  const nodeProvider = web3.getCurrentNodeProvider()
   const txDetails = await nodeProvider.transactions.getTransactionsDetailsTxid(txId, { fromGroup: groupIndex, toGroup: groupIndex })
   const outputIndex = txDetails.unsigned.fixedOutputs.length
   const hex = txId + outputIndex.toString(16).padStart(8, '0')
-  const hashHex = web3.binToHex(blake.blake2b(web3.hexToBinUnsafe(hex), undefined, 32))
+  const hashHex = binToHex(blake.blake2b(hexToBinUnsafe(hex), undefined, 32))
   return hashHex.slice(0, 62) + groupIndex.toString(16).padStart(2, '0')
+}
+
+export interface NFTCollectionBase {
+  id: string
+  name: string
+  description: string
+  owner: string
+  totalSupply: bigint
+  image: string
+  nfts: NFT[]
+}
+
+export interface NFTOpenCollection extends NFTCollectionBase {
+  collectionType: 'NFTOpenCollection'
+}
+
+export interface NFTPublicSaleCollection extends NFTCollectionBase {
+  collectionType: 'NFTPublicSaleCollectionSequential' | 'NFTPublicSaleCollectionRandom'
+  maxSupply: bigint
+  mintPrice: bigint
+  maxBatchMintSize: number
+  nftBaseUri: string
+}
+
+export type NFTOpenCollectionMetadata = Omit<NFTOpenCollection, 'nfts'>
+export type NFTPublicSaleCollectionMetadata = Omit<NFTPublicSaleCollection, 'nfts'>
+
+export type NFTCollection = NFTOpenCollection | NFTPublicSaleCollection
+export type NFTCollectionMetadata = NFTOpenCollectionMetadata | NFTPublicSaleCollectionMetadata
+
+export type NFTsByCollection = Map<NFTCollection, NFT[]>
+
+async function fetchNonEnumerableNFTs(addresses: string[], listed: boolean): Promise<NFT[]> {
+  if (addresses.length === 0) return []
+  const nodeProvider = web3.getCurrentNodeProvider()
+  const methodIndexes = [0, 1] // getTokenUri, getCollectionId
+  const calls = addresses.flatMap((address) => methodIndexes.map((idx) => ({
+    group: groupOfAddress(address),
+    address: address,
+    methodIndex: idx
+  })))
+  const callResult = await nodeProvider.contracts.postContractsMulticallContract({ calls })
+  const getNFT = async (address: string, metadataUri: string, collectionId: string): Promise<NFT | undefined> => {
+    try {
+      if (metadataUri && collectionId) {
+        const metadata = (await axios.get(metadataUri)).data
+        return {
+          name: metadata.name,
+          description: metadata.description,
+          image: metadata.image,
+          tokenId: binToHex(contractIdFromAddress(address)),
+          collectionId: collectionId,
+          minted: true,
+          listed
+        }
+      }
+    } catch (error) {
+      console.error(`failed to get non-enumerable nft, collection id: ${collectionId}, address: ${address} error: ${error}`)
+    }
+    return undefined
+  }
+  const promises = addresses.map((address, idx) => {
+    const callResultIndex = idx * 2
+    const metadataUri = hexToString(callResult.results[callResultIndex].returns[0].value as string)
+    const collectionId = callResult.results[callResultIndex + 1].returns[0].value as string
+    return getNFT(address, metadataUri, collectionId)
+  })
+  return (await Promise.all(promises)).filter((nft) => nft !== undefined) as NFT[]
+}
+
+async function fetchEnumerableNFTs(collectionMetadata: NFTPublicSaleCollectionMetadata, indexes: number[], listed: boolean): Promise<NFT[]> {
+  if (collectionMetadata.nftBaseUri === undefined) return []
+  const groupIndex = groupOfAddress(addressFromContractId(collectionMetadata.id))
+  const getNFT = async (index: number): Promise<NFT | undefined> => {
+    try {
+      const hexStr = collectionMetadata.nftBaseUri + Buffer.from(index.toString(), 'ascii').toString('hex')
+      const tokenUri = hexToString(hexStr)
+      const metadata = (await axios.get(tokenUri)).data
+      const tokenId = subContractId(collectionMetadata.id, binToHex(encodeU256(BigInt(index))), groupIndex)
+      return {
+        name: metadata.name,
+        description: metadata.description,
+        image: metadata.image,
+        tokenId: tokenId,
+        collectionId: collectionMetadata.id,
+        listed: listed,
+        minted: false,
+        tokenIndex: index
+      }
+    } catch (error) {
+      console.error(`failed to fetch enumerable nft, collection id: ${collectionMetadata.id}, index: ${index}, error: ${error}`)
+    }
+    return undefined
+  }
+  const promises = indexes.map((index) => getNFT(index))
+  return (await Promise.all(promises)).filter((nft) => nft !== undefined) as NFT[]
+}
+
+export async function fetchNFTByPage(collectionMetadata: NFTCollectionMetadata, page: number, pageSize: number): Promise<NFT[]> {
+  const skipped = page * pageSize
+  if (collectionMetadata.collectionType === 'NFTOpenCollection') {
+    const explorerProvider = web3.getCurrentExplorerProvider()
+    if (explorerProvider === undefined) return []
+    const collectionAddress = addressFromContractId(collectionMetadata.id)
+    const { subContracts } = await explorerProvider.contracts.getContractsContractSubContracts(collectionAddress)
+    const addresses = (subContracts ?? []).slice(skipped, skipped + pageSize)
+    return await fetchNonEnumerableNFTs(addresses, false)
+  }
+
+  const range = (from: number, count: number): number[] => Array.from(Array(count).keys()).map((v) => from + v)
+  const totalSupply = Number(collectionMetadata.totalSupply)
+  const maxSupply = Number(collectionMetadata.maxSupply!)
+  const indexes = range(skipped, pageSize).filter((idx) => idx < maxSupply)
+  const nfts = await fetchEnumerableNFTs(collectionMetadata, indexes, false)
+  return nfts.map<NFT>((nft) => {
+    if (nft.tokenIndex! < totalSupply) {
+      return {...nft, minted: true}
+    } else {
+      return {...nft, mintPrice: collectionMetadata.mintPrice }
+    }
+  })
+}
+
+// TODO: Improve using multi-call, but it doesn't seem to work for NFTPublicSaleCollection?
+export async function fetchNFTCollectionMetadata(
+  collectionId: string
+): Promise<NFTCollectionMetadata | undefined> {
+  const nodeProvider = web3.getCurrentNodeProvider()
+  const collectionAddress = addressFromContractId(collectionId)
+  const state = await nodeProvider.contracts.getContractsAddressState(collectionAddress, { group: 0 })
+  if (state.codeHash == NFTOpenCollection.contract.codeHash) {
+    const contractState = NFTOpenCollection.contract.fromApiContractState(state) as NFTOpenCollectionTypes.State
+    const metadataUri = hexToString(contractState.fields.collectionUri)
+    const metadata = (await axios.get(metadataUri)).data
+    return {
+      id: collectionId,
+      collectionType: 'NFTOpenCollection',
+      name: metadata.name,
+      description: metadata.description,
+      totalSupply: contractState.fields.totalSupply,
+      owner: contractState.fields.collectionOwner,
+      image: metadata.image
+    }
+  } else if (state.codeHash == NFTPublicSaleCollectionSequential.contract.codeHash) {
+    const contractState = NFTPublicSaleCollectionSequential.contract.fromApiContractState(state) as NFTPublicSaleCollectionSequentialTypes.State
+    const metadataUri = hexToString(contractState.fields.collectionUri)
+    const metadata = (await axios.get(metadataUri)).data
+    return {
+      id: collectionId,
+      collectionType: 'NFTPublicSaleCollectionSequential',
+      name: metadata.name,
+      description: metadata.description,
+      totalSupply: contractState.fields.totalSupply,
+      owner: contractState.fields.collectionOwner,
+      image: metadata.image,
+      maxSupply: contractState.fields.maxSupply,
+      mintPrice: contractState.fields.mintPrice,
+      maxBatchMintSize: Number(contractState.fields.maxBatchMintSize),
+      nftBaseUri: contractState.fields.nftBaseUri
+    }
+  }
 }
